move BBox to request packages
