+++ spi/src/main/java/org/n52/io/request/BBox.java
@@ -26,7 +26,7 @@
  * or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License
  * for more details.
  */
-package org.n52.io.response;
+package org.n52.io.request;
 
 import org.n52.io.geojson.old.GeojsonPoint;
