+++ spi/src/main/java/org/n52/io/request/IoParameters.java
@@ -53,7 +53,6 @@ import org.n52.io.IoParseException;
 import org.n52.io.crs.BoundingBox;
 import org.n52.io.crs.CRSUtils;
 import org.n52.io.geojson.old.GeojsonPoint;
-import org.n52.io.response.BBox;
 import org.n52.io.response.PlatformType;
 import org.n52.io.response.dataset.ValueType;
 import org.n52.io.style.LineStyle;
